added Mutes, Bans and Admins to Chatrooms

+++ app/controllers/chatroom_controller.rb
@@ -10,7 +10,7 @@ class ChatroomController < ApplicationController
 		@chatroom_name = params[:chatroom_name]
 		@chatroom_pw = params[:chatroom_password]
 		if params[:action] != "create"
-			@chatroom = Chatroom.find(params[:id])
+			@chatroom = Chatroom.find(params[:id]) rescue nil
 		end
 	end
 
@@ -18,11 +18,20 @@ class ChatroomController < ApplicationController
 			render json: Chatroom.all_with_subscription_status(@current_user), status: :ok
 	end
 
-	def update # responds to a PATCH/PUT request to "/api/chatroom/:id"
-		if ChatroomMember.find_by(chatroom: @chatroom, user: @current_user) == nil #This user is not yet subscribed to this channel
-			if @chatroom.is_private
-				puts "password is #{@chatroom_pw}, encrypted it is #{Base64.strict_encode64(@chatroom_pw)}"
+	def show # responds to a GET request on "/api/chatroom/:id"
+		puts "in chatroom_controller#show"
+		render json: Chatroom.find(params[:id]), status: :ok
+	end
 
+	def update # responds to a PATCH/PUT request to "/api/chatroom/:id"
+		unless @current_user
+			return render json: { error: "Cannot find current user" }, status: :internal_server_error
+		end
+		if @chatroom.members.find_by(user: @current_user) == nil #This user is not yet subscribed to this channel
+			if @chatroom.bans.find_by(user: @current_user)
+				return render json: { error: "You're not allowed to join this channel, as you've been banned." }, status: :unauthorized
+			end
+			if @chatroom.is_private
 				if @chatroom.password != Base64.strict_encode64(@chatroom_pw)
 					render json: {error: "Wrong password" }, status: :unauthorized
 					return false
@@ -32,25 +41,33 @@ class ChatroomController < ApplicationController
 			if newmember.save
 				@chatroom.amount_members += 1
 				@chatroom.save
-				 ChatroomMember.where(chatroom: @chatroom).each do |m|
-					 if m != newmember
-						 ChatChannel.broadcast_to(m.user, {
-							 title: "groupchat_#{@chatroom.id}",
-							 body: " - #{newmember.user.name} has joined the channel!"
-						 })
-					 end
+				@chatroom.members.each do |m|
+				 if m != newmember
+					 ChatChannel.broadcast_to(m.user, {
+						 title: "groupchat_#{@chatroom.id}",
+						 body: " - #{newmember.user.name} has joined the channel!"
+					 })
 				 end
+				 end
+				render json: { status: "Succesfully joined chatchannel #{@chatroom.name}" }, status: :ok
+			else
+				render json: { error: "Saving new chatroom member failed, sorry" }, status: :unprocessable_entity
 			end
-		end
-		respond_to do |format|
-			format.html { }
-			format.json { head :no_content}
+		else
+			render json: { status: "You've already joined this channel" }, status: :no_content
 		end
 	end
 
 	def destroy # responds to a DELETE request to "/api/chatroom/:id"
-		puts "in chatroom_controller#destroy"
-		myChatroomMember = ChatroomMember.find_by(chatroom: @chatroom, user: @current_user)
+		# I use this pre-built method to leave a channel (not delete the channel)
+		unless @current_user
+			return render json: { error: "Cannot find current user" }, status: :internal_server_error
+		end
+		if @chatroom.owner == @current_user
+			@chatroom.destroy
+			return render json: { status: "Owner left the channel, therefore we disband it" }, status: :ok
+		end
+		myChatroomMember = @chatroom.members.find_by(user: @current_user)
 		if myChatroomMember != nil
 			myChatroomMember.destroy
 			@chatroom.amount_members -= 1
@@ -61,11 +78,10 @@ class ChatroomController < ApplicationController
 		end
 	end
 
-	def show # responds to a GET request on "/api/chatroom/:id"
-		puts "in chatroom_controller#show"
-	end
-
 	def create # responds to a POST request on "/api/chatroom"
+		unless @current_user
+			return render json: { error: "Cannot find current user" }, status: :internal_server_error
+		end
 		if @chatroom_name == nil or @chatroom_name.empty?
 			render json: { error: "You need to specify a valid name" }, status: :bad_request
 			return
